test: typing test
