+++ vscode-extension/test/e2e/defaults-sass/util.js
@@ -4,6 +4,7 @@ const {
 	showFile,
 	sleep,
 	sleepCI,
+	type,
 } = require("../util");
 
 /**
@@ -21,36 +22,6 @@ function getDocPath(p) {
 function getDocUri(p) {
 	return vscode.Uri.file(getDocPath(p));
 }
-
-/**
- * @private
- * @param {import('vscode').TextDocument} doc
- * @returns {Promise<import('vscode').TextDocument>}
- */
-function onDocumentChange(doc) {
-	return new Promise(resolve => {
-		const sub = vscode.workspace.onDidChangeTextDocument(e => {
-			if (e.document !== doc) {
-				return;
-			}
-			sub.dispose();
-			resolve(e.document);
-		});
-	});
-};
-
-/**
- *
- * @param {import('vscode').TextDocument} document
- * @param {string} text
- * @returns {Promise<import('vscode').TextDocument>}
- */
-async function type(document, text) {
-	const onChange = onDocumentChange(document);
-	await vscode.commands.executeCommand('type', { text });
-	await onChange;
-	return document;
-};
 
 module.exports = {
 	sleep,
